Moved pages to src folder
