--- pages/index.tsx
@@ -1,49 +1,0 @@
-import { Grid, GridItem } from "@chakra-ui/react";
-import { GetStaticProps } from "next";
-import Head from "next/head";
-import { initializeApollo } from "../apollo/client";
-import { FeedDocument } from "../apollo/generated/graphql";
-import Layout from "../components/layout/Layout";
-import PostList from "../components/post/PostList";
-import PostsLastWeek from "../components/post/PostsLastWeek";
-import HiringNow from "../components/sidebar/HiringNow";
-import Tags from "../components/sidebar/Tags";
-
-export default function Home() {
-  return (
-    <div>
-      <Head>
-        <title>Hacker Hunt</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Layout>
-        <Grid h="200px" templateColumns="repeat(6, 1fr)" gap={6} mt={6}>
-          <GridItem colSpan={{ base: 6, md: 4 }}>
-            <PostsLastWeek />
-            <PostList />
-          </GridItem>
-          <GridItem colSpan={2} display={{ base: "none", md: "block" }}>
-            <HiringNow />
-            <Tags />
-          </GridItem>
-        </Grid>
-      </Layout>
-    </div>
-  );
-}
-
-export const getStaticProps: GetStaticProps = async () => {
-  const client = await initializeApollo({});
-
-  await client.query({
-    query: FeedDocument,
-  });
-
-  return {
-    revalidate: 60 * 60,
-    props: {
-      initialApolloState: JSON.parse(JSON.stringify(client.cache.extract())),
-    },
-  };
-};
