--- pages/new.tsx
@@ -1,19 +1,0 @@
-import Head from "next/head";
-import CreatePostForm from "../components/forms/CreatePostForm";
-import Layout from "../components/layout/Layout";
-import { withAuth } from "../hoc/withAuth";
-
-export default withAuth(function New() {
-  return (
-    <div>
-      <Head>
-        <title>Hacker Hunt</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Layout>
-        <CreatePostForm />
-      </Layout>
-    </div>
-  );
-});
